requirements.txt
flask
requests
flask-limiter
python-dotenv
razorpay
flask_cors
setuptools
flask_sqlalchemy
pandas